moving variable to other place so that during application startup it doesn't try to load testing requirements
