+++ stubo/service/__init__.py
@@ -7,3 +7,6 @@
     :copyright: (c) 2015 by OpenCredo.
     :license: GPLv3, see LICENSE for more details.
 """
+# defining DummyModel, this variable is overwritten during tests
+from tornado.util import ObjectDict
+DummyModel = ObjectDict
